requirements.txt
django
psycopg2-binary
dj-database-url
python-dotenv
gunicorn
whitenoise
django-cors-headers
djangorestframework